Add docs component for our /open-source site.
