+++ package.json
@@ -27,25 +27,25 @@
   },
   "dependencies": {
     "builder": "~2.9.1",
-    "builder-victory-component": "^2.1.7",
+    "builder-victory-component": "^2.1.8",
     "radium": "^0.17.1",
     "victory-chart": "^9.1.2",
     "victory-pie": "^4.0.0"
   },
   "devDependencies": {
-    "builder-victory-component-dev": "^2.1.7",
+    "builder-victory-component-dev": "^2.1.8",
     "chai": "^3.2.0",
     "formidable-landers": "^2.0.0",
     "history": "~1.17.0",
     "lodash": "^4.12.0",
     "mocha": "^2.2.5",
     "opener": "^1.4.1",
-    "react-router": "^2.4.0",
     "react": "~15.1.0",
     "react-addons-test-utils": "~15.1.0",
     "react-dom": "~15.1.0",
     "react-hot-loader": "^1.2.8",
-    "scroll-behavior": "^0.3.3",
+    "react-router": "^2.4.0",
+    "react-router-scroll": "^0.2.0",
     "sinon": "^1.16.1",
     "sinon-chai": "^2.8.0"
   }
